HDX-9781 Move fund_view
