+++ src/hapi_schema/db_funding.py
@@ -5,9 +5,7 @@ from datetime import datetime
 from sqlalchemy import (
     CheckConstraint,
     DateTime,
-    Float,
     ForeignKey,
-    Integer,
     String,
     select,
     text,
@@ -93,21 +91,3 @@ view_params_funding = ViewParams(
         )
     ),
 )
-
-
-class DBFundingVAT(Base):
-    __tablename__ = "funding_vat"
-    resource_hdx_id: Mapped[str] = mapped_column(String(36))
-    appeal_code: Mapped[str] = mapped_column(String(32), primary_key=True)
-    location_ref: Mapped[int] = mapped_column(Integer, primary_key=True)
-    appeal_name: Mapped[str] = mapped_column(String(256))
-    appeal_type: Mapped[str] = mapped_column(String(32))
-    requirements_usd: Mapped[float] = mapped_column(Float, index=True)
-    funding_usd: Mapped[float] = mapped_column(Float, index=True)
-    funding_pct: Mapped[float] = mapped_column(Float, index=True)
-    reference_period_start: Mapped[datetime] = mapped_column(
-        DateTime, index=True
-    )
-    reference_period_end: Mapped[datetime] = mapped_column(DateTime)
-    location_code: Mapped[str] = mapped_column(String(128), index=True)
-    location_name: Mapped[str] = mapped_column(String(512), index=True)
