+++ src/hapi_schema/db_views_as_tables.py
@@ -84,3 +84,23 @@ class DBFoodSecurityVAT(Base):
     admin2_name: Mapped[str] = mapped_column(String(512), index=True)
     admin2_is_unspecified: Mapped[bool] = mapped_column(Boolean)
     admin1_ref: Mapped[int] = mapped_column(Integer)
+
+
+class DBFundingVAT(Base):
+    __tablename__ = "funding_vat"
+    resource_hdx_id: Mapped[str] = mapped_column(String(36))
+    appeal_code: Mapped[str] = mapped_column(String(32), primary_key=True)
+    location_ref: Mapped[int] = mapped_column(Integer, primary_key=True)
+    appeal_name: Mapped[str] = mapped_column(String(256))
+    appeal_type: Mapped[str] = mapped_column(String(32))
+    requirements_usd: Mapped[float] = mapped_column(Float, index=True)
+    funding_usd: Mapped[float] = mapped_column(Float, index=True)
+    funding_pct: Mapped[float] = mapped_column(Float, index=True)
+    reference_period_start: Mapped[datetime] = mapped_column(
+        DateTime, index=True
+    )
+    reference_period_end: Mapped[datetime] = mapped_column(
+        DateTime, nullable=True
+    )
+    location_code: Mapped[str] = mapped_column(String(128), index=True)
+    location_name: Mapped[str] = mapped_column(String(512), index=True)
