+++ tests/test_funding.py
@@ -29,24 +29,6 @@ def test_funding_view(run_view_test):
     )
 
 
-def test_food_security_vat(
-    run_indexes_test, run_columns_test, run_primary_keys_test
-):
-    """Check that the food_security view as table is correct - columns match, expected indexes present"""
-    expected_primary_keys = ["appeal_code", "location_ref"]
-    expected_indexes = [
-        "requirements_usd",
-        "funding_usd",
-        "funding_pct",
-        "reference_period_start",
-        "location_code",
-        "location_name",
-    ]
-    run_columns_test("funding_vat", "funding_view", view_params_funding)
-    run_indexes_test("funding_vat", expected_indexes)
-    run_primary_keys_test("funding_vat", expected_primary_keys)
-
-
 def test_reference_period_constraint(run_constraints_test):
     """Check that reference_period_end cannot be less than start"""
     run_constraints_test(
